requirements.txt
flask==2.0.1
requests==2.25.1
pyyaml==5.4.1
jinja2==3.0.1
werkzeug==2.0.1
cryptography==3.4.7
pillow==8.3.1
paramiko==2.7.2
urllib3==1.26.5
pyjwt==2.1.0
sqlalchemy==1.4.22
lxml==4.6.3
numpy==1.21.0
setuptools==57.0.0
certifi==2021.5.30
